+++ tests/Frameworks/Laravel/Version_5_8/routes/web.php
@@ -10,7 +10,6 @@
 | contains the "web" middleware group. Now create something great!
 |
 */
-
-Route::get('/', function () {
-    return view('welcome');
-});
+Route::get('/simple', ['as' => 'simple_route', 'uses' => 'CommonSpecsController@simple']);
+Route::get('/simple_view', 'CommonSpecsController@simple_view');
+Route::get('/error', 'CommonSpecsController@error');