+++ tests/Integrations/Laravel/V5_7/PipelineTracingTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DDTrace\Tests\Integrations\Laravel\V5;
+namespace DDTrace\Tests\Integrations\Laravel\V5_7;
 
 use DDTrace\Tests\Common\SpanAssertion;
 use DDTrace\Tests\Common\SpanAssertionTrait;