+++ tests/Integrations/Laravel/V5_8/CommonScenariosTest.php
@@ -1,0 +1,93 @@
+<?php
+
+namespace DDTrace\Tests\Integrations\Laravel\V5_8;
+
+use DDTrace\Tests\Common\SpanAssertion;
+use DDTrace\Tests\Common\WebFrameworkTestCase;
+use DDTrace\Tests\Frameworks\Util\Request\RequestSpec;
+
+final class CommonScenariosTest extends WebFrameworkTestCase
+{
+    protected static function getAppIndexScript()
+    {
+        return __DIR__ . '/../../../Frameworks/Laravel/Version_5_8/public/index.php';
+    }
+
+    protected static function getEnvs()
+    {
+        return [
+            'APP_NAME' => 'laravel_test_app',
+        ];
+    }
+
+    /**
+     * @dataProvider provideSpecs
+     * @param RequestSpec $spec
+     * @param array $spanExpectations
+     * @throws \Exception
+     */
+    public function testScenario(RequestSpec $spec, array $spanExpectations)
+    {
+        $traces = $this->tracesFromWebRequest(function () use ($spec) {
+            $this->call($spec);
+        });
+
+        $this->assertExpectedSpans($this, $traces, $spanExpectations);
+    }
+
+    public function provideSpecs()
+    {
+        return $this->buildDataProvider(
+            [
+                'A simple GET request returning a string' => [
+                    SpanAssertion::build(
+                        'laravel.request',
+                        'laravel_test_app',
+                        'web',
+                        'App\Http\Controllers\CommonSpecsController@simple simple_route'
+                    )->withExactTags([
+                        'laravel.route.name' => 'simple_route',
+                        'laravel.route.action' => 'App\Http\Controllers\CommonSpecsController@simple',
+                        'http.method' => 'GET',
+                        'http.url' => 'http://localhost:9999/simple',
+                        'http.status_code' => '200',
+                    ]),
+                ],
+                'A simple GET request with a view' => [
+                    SpanAssertion::build(
+                        'laravel.request',
+                        'laravel_test_app',
+                        'web',
+                        'App\Http\Controllers\CommonSpecsController@simple_view unnamed_route'
+                    )->withExactTags([
+                        'laravel.route.action' => 'App\Http\Controllers\CommonSpecsController@simple_view',
+                        'http.method' => 'GET',
+                        'http.url' => 'http://localhost:9999/simple_view',
+                        'http.status_code' => '200',
+                    ])->withExistingTagsNames(['laravel.route.name']),
+                    SpanAssertion::build(
+                        'laravel.view',
+                        'laravel_test_app',
+                        'web',
+                        'laravel.view'
+                    ),
+                ],
+                'A GET request with an exception' => [
+                    SpanAssertion::build(
+                        'laravel.request',
+                        'laravel_test_app',
+                        'web',
+                        'App\Http\Controllers\CommonSpecsController@error unnamed_route'
+                    )->withExactTags([
+                        'laravel.route.name' => '',
+                        'laravel.route.action' => 'App\Http\Controllers\CommonSpecsController@error',
+                        'http.method' => 'GET',
+                        'http.url' => 'http://localhost:9999/error',
+                        'http.status_code' => '500',
+                    ]),
+                    SpanAssertion::exists('laravel.view')
+                ],
+            ]
+        );
+    }
+}